--- Game/Model/Fight.py
@@ -1,12 +1,0 @@
-class Fight:
-
-    def __init__(self, id=None, gold=None, xp=None, enemies=None, player=None):
-        self.__id = id
-        self.__gold = gold
-        self.__xp = xp
-        self.__enemies = enemies
-        self.__player = player
-
-    def Start(self):
-        return
-
