+++ main.py
@@ -1,42 +1,17 @@
 # main du projet Chivalery
+import pygame
+from Game.Menu.Application import *
 
-# import pygame
-# from Game.Menu.Application import *
-#
-# pygame.init()
-#
-#
-# app = Application()
-# app.menu()
-#
-# clock = pygame.time.Clock()
-#
-# while app.statut:
-#     app.update()
-#     clock.tick(30)
-#
-# pygame.quit()
+pygame.init()
 
-from Game.Model.Knight import Knight
-from Game.Model.KnightClasse import KnightClasse
-from Game.Model.Level import Level
 
-# KnightClasse_id = 1
-# KnightClasse_label = "ecu"
-# KnightClasse_speciality = ""
-# KnightClasse_modifierAttack = 0
-# KnightClasse_modifierDefense = 0
-# KnightClasse_modifierSpeciality = 0
-# KnightClasse_listSkill = []
-#
-# k1 = Knight(0, "arthur", "0")
-#
-# lvl = Level()
-# print(lvl.getExpLvlByLvl(1))
-# print(lvl.getExpLvlByLvl(2))
-# print(lvl.getExpLvlByLvl(3))
+app = Application()
+app.menu()
 
-id=None
-print(id)
+clock = pygame.time.Clock()
 
-# k1.print()
+while app.statut:
+    app.update()
+    clock.tick(30)
+
+pygame.quit()
